backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flasgger==0.9.7.1
anthropic==0.34.2
groq==0.11.0
python-dotenv==1.0.1
Pillow==10.4.0
sqlalchemy==2.0.35
tensorflow==2.20.0
numpy==1.26.4